Attendance Summary(전체 근무 일, 시간, 잔여 휴가, 지각 횟수) / Attendance weekly 근무시간 progressbar 생성

+++ src/main/java/org/goodee/startup_BE/attendance/dto/AttendanceResponseDTO.java
@@ -30,10 +30,10 @@ public class AttendanceResponseDTO {
     private LocalDate attendanceDate;
 
     @Schema(description = "직원 연차 개수", example = "15")
-    private Integer annualLeaveCount;
+    private Long annualLeaveCount;
 
     @Schema(description = "총 근무 일수", example = "30일")
-    private Integer workDate;
+    private Long workDate;
 
     @Schema(description = "출근 시간", example = "2025-10-27T09:00:00")
     private LocalDateTime startTime;
